requirements.txt
flet>=0.21.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0 